SPARKC-330: Refactor CassandraRDDPartitioner to have static Partitons
In order to benefit from implementing org.apache.spark.Partitioner we
need a given partitioner to be locked into a set number of partitions
based on it's constructor. This means we can't allow the lookup to be a
function of parameters like the given "whereClause" in a lookup. I've
moved that logic back out to the CassandraTableScanRDD. This will
provide room for the next move of sharing a partitioner between
TableScan RDDs.