SPARKC-330: Refactor Partitioner out of CassandraPartition Generator
To better divide responsibilites CassandraRDDPartitioner is renamed
CassandraRDDPartitionerGenerator and a new CassandraRDDPartitioner
implementing org.apache.spark.Partitioner is created.

The new partitoner class uses a new type parameter for determining the
partition key of the given RDD. This is used to generate a RowWriter
which will use the incoming Key type to take scala values and bind them
into a prepared statement. The Token Generator uses this to determine
the underlying C* token and returns it.

Currently this is uncoupled from the CTablescanRDD.