SPARKC-330: Modify KeyBy methods to automatically assign a Partitioner
Assigns a Partitioner to a PairRDD when created by KeyBy and the Key
type contains the complete partition key (in whatever form). This is
accomplished by changing the implicit parameters on keyBy to also
include a RowWriter for the Key Type. This is used along with the
current partitioning to generate a Partitioner which can map keys of
type K back to the Token which C* will have stored it with.

This change is a bit more dramatic on the Java side as all the keyBy
methods now also require specifying the RowWriterFactory.