SPARKC-330: Determine which Partition Contains a Token
An ordering is imposed on our Tokens so they can be compared with the
non-serializable Java Driver Tokens. We now can determine which
Cassandra Partition contains a Token by searching through the given
partitions one at a time and seeing which partition a token should fall
into.

In addition the tests are now bolstered to actually test that the
partitioning is correctly recovered.