SPARKC-330: Adds real join tests, Fixes partitioner serialization
Adds a suite of tests to make sure that CassandraPartitioner works
correctly when serialized for a join. This required fixing the
partitioner so that the TokenGenerator class is not serialized but is
instead created after serialization. The same trick is not possible with
the rowWriter class as it must be made prior to serilization and
serialized.