SPARKC-330: Adds keyByAndApply For Java Functionality
Since the JavaAPI does not allow us to specify implicits and it loses
the TableScan type upon using keyBy, a new approach was needed for
applying the partitioner from one RDD to another.

keyByAndApply will call both keyBy and then applyPartitionerFrom in
order. This function can be called from a CassandraTableScanJavaRDD
circumventing the PairRDD created from the keyBy operation.