SPARKC-330: Allow Partitioning of keys with Different Names
In order to allow tables like (x int key, a int, b int) to join against
                  tables like (y int key, c int, d int)
the CassandraPartitioner is changed to measure equivelence only on the
underlying token ranges used for comparison and the given keyspace name.

Now when applying a new Partitioner to an RDD, if there is an existing
RDD, it is left intact but it's partitioning and token range are
replaced. This allows for mixing different RowWriters as long as they
are of the same Type even if they map to different underlying Cassandra
Tables.

To allow the switiching of the partition key column names the
CqlTokenRange now holds a TokenClause whose partition key can change.