SPARKC-330: Refactor to allow for KeyMapping in Partitioner
To allow the key mapper to be used by the partitioner several refactors
are made to the CassandraRDDPartitioner.

First the CassandraRDDPartitioner now only takes two parameters.
A PartitioningMetadata object which contains the interal representation
of the token groupings in the cluster. The tableDef they are defined for
and methods to dynaimcally create the groupings which are made into
CassandraPartitions.

The second parameter is a mapping for the Partitioner which will allow
for arbitrary columns to be mapped to any type K.

To make it easier to fluently modify the Partitioner, the RowWriter is
now made lazily at the time that the RDD has to actually figure out the
partitioing. This lets us modify the tableDef and columnSelector without
having every step along the way have a valid RowWriter.